refactor: move utility funcs.
